requirements.txt
certifi==2023.5.7
chardet==3.0.4
customtkinter==5.1.3
darkdetect==0.8.0
deepl-cli==0.5.0
googletrans==3.1.0a0
greenlet==2.0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
install-playwright==0.0.0
mss==9.0.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
playwright==1.33.0
pyee==9.0.4
pytesseract==0.3.10
rfc3986==1.5.0
sniffio==1.3.0
typing_extensions==4.6.0
